refactor: update layouts for dropdown versions
