+++ .vscode/settings.json
@@ -4,6 +4,7 @@
         "autocorrect",
         "bibtemplate",
         "borderless",
+        "boxshadow",
         "csspart",
         "DDMM",
         "describedby",
