+++ components/dropdown/apiExamples/emphasized/basic.html
@@ -1,4 +1,4 @@
-<auro-dropdown aria-label="custom label" shape="pill" size="xl" layout="emphasized" style="width: 249px;">
+<auro-dropdown aria-label="custom label" shape="pill" size="xl" layout="emphasized" style="width: 249px;" ondark>
   Lorem ipsum solar
   <div slot="trigger">
     Trigger
@@ -8,7 +8,7 @@
   </span>
 </auro-dropdown>
 
-<auro-dropdown aria-label="custom label" shape="pill-left" size="xl" layout="emphasized" style="width: 249px;">
+<auro-dropdown aria-label="custom label" shape="pill-left" size="xl" layout="emphasized" style="width: 249px;" ondark>
   Lorem ipsum solar
   <div slot="trigger">
     Trigger
@@ -18,7 +18,7 @@
   </span>
 </auro-dropdown>
 
-<auro-dropdown aria-label="custom label" shape="pill-right" size="xl" layout="emphasized" style="width: 249px;">
+<auro-dropdown aria-label="custom label" shape="pill-right" size="xl" layout="emphasized" style="width: 249px;" ondark>
   Lorem ipsum solar
   <div slot="trigger">
     Trigger
