+++ components/dropdown/docs/partials/index.md
@@ -27,8 +27,15 @@ Not including one of the above options will result in your UI being non-complian
 
 ## Emphasized Layouts
 
-<div class="exampleWrapper">
+<div class="exampleWrapper--ondark">
   <!-- AURO-GENERATED-CONTENT:START (FILE:src=./../apiExamples/emphasized/basic.html) -->
+  <!-- AURO-GENERATED-CONTENT:END -->
+</div>
+
+## Snowflake Layouts
+
+<div class="exampleWrapper--ondark">
+  <!-- AURO-GENERATED-CONTENT:START (FILE:src=./../apiExamples/snowflake/basic.html) -->
   <!-- AURO-GENERATED-CONTENT:END -->
 </div>
