+++ components/input/src/styles/emphasized/style.scss
@@ -141,6 +141,11 @@ input {
     justify-content: flex-start;
   }
 
+  // HIDE BORDER WHEN NOT FOCUSED AND NOT INVALID
+  &:not(:focus-within):not(:is([validity]:not([validity='valid']))) {
+    --ds-auro-input-border-color: transparent;
+  }
+
   &:focus-within {
     @include emphasizedActiveLabel;
     @include emphasizedActiveInput;
